requirements.txt
discord.py>=2.3.2
yt-dlp>=2023.12.30
aiohttp>=3.9.1
aiofiles>=23.2.1
Pillow>=10.1.0
python-dotenv>=1.0.0
asyncio-throttle>=1.0.2
beautifulsoup4>=4.12.0
lxml>=4.9.0
PyNaCl>=1.5.0
